requirements.txt
# =============================================================================
# 核心必需依赖 (Core Required Dependencies)
# =============================================================================

# 医学图像处理
SimpleITK>=2.0.0              # 医学图像I/O和处理
numpy>=1.20.0                 # 数值计算基础
scipy>=1.7.0                  # 科学计算 (distance, ndimage)

# 图像处理和机器学习
scikit-image>=0.18.0          # 图像处理 (skeletonize_3d)
scikit-learn>=1.0.0           # 机器学习工具 (train_test_split, metrics)
networkx>=2.6.0               # 图论算法 (MST, DFS)

# PyTorch 生态系统
torch>=1.10.0                 # 深度学习框架
torch-geometric>=2.0.0        # 图神经网络
torch-scatter                 # torch-geometric依赖

# 训练基础设施
tqdm>=4.62.0                  # 进度条
tensorboard>=2.7.0            # 训练监控

# Python标准库增强
PyYAML>=5.4.0                # YAML配置文件支持

# =============================================================================
# 可选依赖 (Optional Dependencies)
# =============================================================================

# 可视化和分析 (用于结果分析，可选)
matplotlib>=3.5.0             # 基础绘图
seaborn>=0.11.0              # 统计图表 (混淆矩阵等)

# 3D可视化 (高级功能，可选)
# vtk>=9.0.0                  # 3D可视化 (如需要请取消注释)

# =============================================================================
# 注意事项
# =============================================================================
# 1. torchvision 在当前项目中未使用，已移除
# 2. 请根据您的CUDA版本安装对应的PyTorch版本
# 3. torch-scatter 需要单独安装: pip install torch-scatter -f https://data.pyg.org/whl/torch-1.10.0+cu113.html
# 4. VTK体积较大，仅在需要3D可视化时安装